Fix crash after replay of DEALLOCATE ALL
Replaying DEALLOCATE ALL executes "remove_all_pstmts", which
failed to set the list of prepared statements to NULL after
freeing it.  This causes the next execution of that function
for the same connection to fail, for example when the end of
the session is replayed.

This closes #23, reported by Wan Shen Lim.